# docs/openapi/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "support@swagger.io",
      "name" : "API Support",
      "url" : "http://www.swagger.io/support"
    },
    "description" : "LocalPaas Backend",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "LocalPaas Backend",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "/v1"
  } ],
  "paths" : {
...
    "/auth/login-options" : {
      "get" : {
        "description" : "Gets login options",
        "operationId" : "getLoginOptions",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.GetLoginOptionsResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Gets login options",
        "tags" : [ "sessions" ]
      }
    },
    "/auth/login-with-api-key" : {
      "post" : {
        "description" : "Login to system with API key",
        "operationId" : "loginWithAPIKey",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithAPIKeyReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithAPIKeyResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Login to system with API key",
        "tags" : [ "sessions" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/auth/login-with-passcode" : {
      "post" : {
        "description" : "Login to system with passcode after using password",
        "operationId" : "loginWithPasscode",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasscodeReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasscodeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Login to system with passcode after using password",
        "tags" : [ "sessions" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/auth/login-with-password" : {
      "post" : {
        "description" : "When you get response's `next_step` with value `NextMfa`, you need to call the API\n`/login-with-passcode` to complete the login process.",
        "operationId" : "loginWithPassword",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasswordReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasswordResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Login to system with username/password",
        "tags" : [ "sessions" ],
        "x-codegen-request-body-name" : "body"
      }
    },
...
    "/internal/auth/dev-mode-login" : {
      "post" : {
        "description" : "Login user for dev mode. `user_id` params is required.",
        "operationId" : "devModeLogin",
        "parameters" : [ {
          "description" : "user id to login",
          "in" : "query",
          "name" : "user_id",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DevModeLoginResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Login user for dev mode",
        "tags" : [ "sessions" ]
      }
    },
...
    "/sessions" : {
      "delete" : {
        "description" : "Deletes the current user session",
        "operationId" : "deleteSession",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DeleteSessionResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Deletes the current user session",
        "tags" : [ "sessions" ]
      }
    },
    "/sessions/delete-all" : {
      "post" : {
        "description" : "Deletes all sessions of the user",
        "operationId" : "deleteAllSessions",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DeleteAllSessionsResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Deletes all sessions of the user",
        "tags" : [ "sessions" ]
      }
    },
...
    "/sessions/refresh" : {
      "post" : {
        "description" : "Refreshes the current user session. Refresh token is required via either\n`Authorization` header or `refresh_token` cookie.",
        "operationId" : "refreshSession",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.RefreshSessionResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Refreshes the current user session",
        "tags" : [ "sessions" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo" : {
        "properties" : {
          "cause" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "debugLog" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.InnerErrorInfo"
            },
            "type" : "array"
          },
          "stackTrace" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "detail", "status", "title", "type" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_apperrors.InnerErrorInfo" : {
        "properties" : {
          "cause" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message", "path" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_base.MFAType" : {
        "enum" : [ "totp", "email" ],
        "type" : "string",
        "x-enum-varnames" : [ "MFATypeTOTP", "MFATypeEmail" ]
      },
...
      "github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.BaseCreateSessionResp" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "accessTokenExp" : {
            "type" : "string"
          }
        },
        "required" : [ "accessToken", "accessTokenExp" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DeleteAllSessionsResp" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DeleteSessionResp" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DevModeLoginDataResp" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "accessTokenExp" : {
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "refreshTokenExp" : {
            "type" : "string"
          }
        },
        "required" : [ "accessToken", "accessTokenExp", "refreshToken", "refreshTokenExp" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DevModeLoginResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DevModeLoginDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.GetLoginOptionsResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginOptionsResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginOptionsResp" : {
        "properties" : {
          "allowLoginWithGitHub" : {
            "type" : "boolean"
          },
          "allowLoginWithGitLab" : {
            "type" : "boolean"
          }
        },
        "required" : [ "allowLoginWithGitHub", "allowLoginWithGitLab" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithAPIKeyDataResp" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "accessTokenExp" : {
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "refreshTokenExp" : {
            "type" : "string"
          }
        },
        "required" : [ "accessToken", "accessTokenExp", "refreshToken", "refreshTokenExp" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithAPIKeyReq" : {
        "properties" : {
          "keyId" : {
            "type" : "string"
          },
          "secretKey" : {
            "type" : "string"
          }
        },
        "required" : [ "keyId", "secretKey" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithAPIKeyResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithAPIKeyDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasscodeDataResp" : {
        "properties" : {
          "session" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.BaseCreateSessionResp"
          }
        },
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasscodeReq" : {
        "properties" : {
          "mfaToken" : {
            "type" : "string"
          },
          "passcode" : {
            "type" : "string"
          }
        },
        "required" : [ "mfaToken", "passcode" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasscodeResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasscodeDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasswordDataResp" : {
        "properties" : {
          "mfaToken" : {
            "type" : "string"
          },
          "mfaType" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.MFAType"
          },
          "nextStep" : {
            "type" : "string"
          },
          "session" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.BaseCreateSessionResp"
          }
        },
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasswordReq" : {
        "properties" : {
          "password" : {
            "type" : "string"
          },
          "trustedDeviceId" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "required" : [ "password", "trustedDeviceId", "username" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasswordResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasswordDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.RefreshSessionDataResp" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "accessTokenExp" : {
            "type" : "string"
          }
        },
        "required" : [ "accessToken", "accessTokenExp" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.RefreshSessionResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.RefreshSessionDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "scheme" : "basic",
        "type" : "http"
      }
    }
  },
...
}